% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trajectories.R
\name{build_trajectories}
\alias{build_trajectories}
\title{Build table of parameter trajectory estimates}
\usage{
build_trajectories(
  param,
  infection_days,
  fit,
  nsim = 1000,
  jurisdictions = NULL
)
}
\arguments{
\item{param}{model parameter of interest}

\item{infection_days}{full date sequence of infection timeseries}

\item{fit}{model fit}

\item{nsim}{number of simulations from model - must be <= samples taken}

\item{jurisdictions}{optional vector of jurisdictions}
}
\value{
table with jurisdiction, date, and all draws for given parameter
}
\description{
Build an output table with all the simulations in order
to pass along to the decision support engine.
}
