% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence}
\usage{
check_convergence(draws, max_r_hat = 1.1, min_n_eff = 1000)
}
\arguments{
\item{draws}{fitted draws}

\item{max_r_hat}{maximum acceptable r hat}

\item{min_n_eff}{minimum acceptable n effective}
}
\value{
TRUE or FALSE on whether convergence was achieved per parameters
}
\description{
Check whether the sampler has converged to our standards.
}
