% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_constant_value.R
\name{expand_constant_value}
\alias{expand_constant_value}
\title{Expand constant value into long tibble}
\usage{
expand_constant_value(dates, jurisdictions, constant_val, col_name)
}
\arguments{
\item{dates}{infection dates sequence}

\item{jurisdictions}{jurisdiction names}

\item{constant_val}{value to be replicated in each cell}

\item{col_name}{column name to contain the \verb{constant value}}
}
\value{
long tibble with constant value replicated
}
\description{
The lowerGPreff model functions expect data in a long format,
which is structed to have a value for every unique date and jurisdiction
pair. This function create a tibble of this structure out of a single
value that should be replicated in each cell.
}
