% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(
  model,
  n_chains = 4,
  max_convergence_tries = 5,
  warmup = 1000,
  n_samples = 1000,
  n_extra_samples = 1000
)
}
\arguments{
\item{model}{greta model object}

\item{n_chains}{number of chains}

\item{max_convergence_tries}{maximum number of times extra samples should be
drawn if convergence isn't reached}

\item{warmup}{number of warmup samples}

\item{n_samples}{number of additional samples after warmup}

\item{n_extra_samples}{number of samples to be taken with \code{extra_samples()}}
}
\value{
fitted draws
}
\description{
Fit model
}
