% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_convergence_stats}
\alias{get_convergence_stats}
\title{Get convergence stats}
\usage{
get_convergence_stats(draws)
}
\arguments{
\item{draws}{fitted draws}
}
\value{
invisibly return list of max r hat and min n effective
}
\description{
Get convergence stats
}
