% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_sims.R
\name{plot_timeseries_sims}
\alias{plot_timeseries_sims}
\title{Plot timeseries simulations}
\usage{
plot_timeseries_sims(
  filename,
  simulations = NULL,
  type = c("notification", "infection", "reff"),
  dates,
  states,
  base_colour = grey(0.4),
  start_date = max(dates) - lubridate::dmonths(1),
  end_date = max(dates),
  case_validation_data = NULL,
  infection_nowcast = TRUE,
  case_forecast = FALSE,
  valid_mat = NULL,
  nowcast_start = NULL,
  dim_sim = c("1", "2"),
  reff_ylim = c(0, 2)
)
}
\arguments{
\item{filename}{output file name}

\item{simulations}{simulations from greta model}

\item{type}{select between notification, infection, and reff}

\item{dates}{infection date sequence}

\item{states}{vector of jurisdictions}

\item{base_colour}{underlying color}

\item{start_date}{first date on x axis}

\item{end_date}{final date on x axis}

\item{case_validation_data}{case data to optionally overlay}

\item{infection_nowcast}{logical indicating whether we should annotate
nowcast horizon}

\item{case_forecast}{logical whether we should forecast cases}

\item{valid_mat}{matrix indicating which case data are valid}

\item{nowcast_start}{start date for beginning of nowcast window}

\item{dim_sim}{whether simulations are in one or two dimensions}

\item{reff_ylim}{limits for y axis on reff plot}
}
\value{
simulation plot
}
\description{
Takes in calculated values from timeseries samples, plots ribbon
plot, adds dates and states labels, and optionally adds in observed data
as overlaying graphical elements.
}
